Add API definition, architecture documents
Perhaps the most important aspect of any given consumable service is its
API. This commit defines the API, using some experimental approaches and
before any other content.

== Design Notes ==

=== The inherent difficulty in expressing "search" behaviour ===

In terms of HTTP REST implementations, search is not quite a trivial one
to implement.

It can be arbitrary complex, filtering down large sets of data to a
limited smaller set of components making it a good candidate for HTTP
methods that traditionally have bodies, such as HTTP POST.

However, POST is designed for requests that have state changes on the
server and cannot be cached -- not necessarily the case for search
results.

Indeed, this use case is so well known that there is an outstanding
proposal to add a new HTTP verb to the existing HTTP verbs that allows
this exact feature: HTTP QUERY

  https://www.w3.org/2012/ldp/wiki/Proposal_for_HTTP_QUERY_Verb

At this time, it does not appear to have gained traction, and the
working group that proposed the QUERY verb is now closed:

  https://www.w3.org/2012/ldp/wiki/Main_Page

in this case, the author opted for the "query" parameter, with other
parameters being stored in complex data (JSON) within the URL. This is
the structure recommended by OpenAPI.

=== Designing the API first ===

Much like code, APIs are going to be consumed far more often than
they're going to be written. However, in the authors experience the API
traditionally tends to be implemented based on the expense to implement
it in the software, rather than it's "consumability".

By designing the API first in a data format and without regard to the
difficulty of implementation the author agrees to a contract as to how
the API will behave without regard for the consequences in terms of
implementation time.

While this may make implementation significantly harder, the hope is
that it will make consumption much easier.

=== Search as a problem ===

search is interesting because it doesn't apply "what" it's searching for.

Consider the following search paths:

  "Wedding dress"

  Can be broken down into its consistent components:

  - "Dress" → a category of clothing
  - "Wedding" → a nebulous concept

  Wedding can mean any number of things; training a model to recognize
  something that's inherently "weddingy" would be an interesting challenge.
  However, for now let's assume we've already trained the ML model on determining which
  social event clothes are appropriate for and tagged the clothes
  appropriately. Other examples include:

  - Cocktail[+Party]
  - Engagement

  Additional concepts beyond the social are an environment:

  - Church
  - Outdoor
  - Beach

  Culturally appropriate clothing:

  - Indian
  - Christian

  There are even first class primitives for clothes that have a
  social context:

  - "Semi Formal"
  - "Formal"
  - "Smart Casual"
  - "Black Tie"

  Accordingly, a search term may convey several possible fields without
  an inherent structure to them that makes them easy to parse.

  Consider: Semi Formal Outdoor Garden Dress

  "Semi Formal" → A class of clothing
  "Outdoor" → A property of clothing implying warmth, hardiness, ability to look clean
  "Garden" → An environment that has a floral dress code
  "Dress" → A shape of clothing

  These things are space delimited except for the "semi formal", of
  which "semi" is a prefix.

Search needs some ability to query multiple fiends with context.

=== Schema.org APIs ===

As an experiment the author has chosen not to define the primitives
themselves, but rather re-use existing primitives from schema.org.

These primitives are somewhat opaque and verbose, but area already in
use in millions of sites around the world -- if not directly as a
consumable HTTP REST API.

It's perhaps reasonable to try and use the primitives provided here, and
if they're unreasonable, work with the schema.org working group to
improve them rather than rewriting them anew.

== References ==

References
 - https://stackoverflow.com/questions/5020704/how-to-design-restful-search-filtering
 - https://www.moesif.com/blog/technical/api-design/REST-API-Design-Filtering-Sorting-and-Pagination/
 - https://swagger.io/docs/specification/serialization/
 - https://stackoverflow.com/questions/3198492/rest-standard-path-parameters-or-request-parameters
 - https://github.com/OAI/OpenAPI-Specification/blob/master/examples/v3.0/uspto.yaml
 - https://softwareengineering.stackexchange.com/questions/233164/how-do-searches-fit-into-a-restful-interface
 - https://magento.stackexchange.com/questions/154963/magento-2-what-are-the-native-required-attributes-to-customer-and-products
 - https://github.com/schemaorg/schemaorg/issues/1685