Add Laravel's lumen to the repository
In order to build a PHP service it's easiest to start with some sort of
PHP framework or application. This commit adds PHP's Lumen application
(a subset of Laravel), which is fairly modern, seems well designed and
in fairly wide use.

There have been no code changes beyond:

  $ lumen new server

== Design Notes ==

=== File Path ===

Ideally, this project will delivery both a server and a sample client
for this application. This commit includes the server codebase; the
client codebase will come later.

It's nested under the "app" directory as that's where all programs go.