Add the Status entity
This commit adds the first of the entities defined in the API to
Lumen.

They're added under an "entities" folder, as Lumen did not appear to
have any canonical convention for housing this non ORM material.

== Design Notes ==

=== Immutable objects ===

The objects created are immutable as the author considers this a
superior conceptual model for entity handling, and there's no reason why
states would need to transition apart from piecemeal object
construction.

=== Chained setters ===

The private "setFoo" methods in the entity object all return `$this` so
they can be chained.

This is simply for convenience.