Extend exception handler to return Status object for method not allowed
The design of the API calls for a clear error return if the implementer
calls an API method that is not supported, such as POST on a controller
that only supports GET.

Laravel implements its controllers by passing only the appropriate HTTP
verb on to the controller, and deferring all else to the exception
handler.

This commit resolves this by extending the exception handler to return
the exception in the case the exception is as a result of the method not
being handled by the router.