Add a test for the invalid method for `/v1/search`
Traditionally the author will implement E2E testing using something like
Dredd. However, in this case it appears that Lumen provides superior
testing primitives and it is trivial to stub the request and directly
query the response data.

This commit uses this to add a test for method not yet implemented.