Add Robofile as a task runner
The author has considered many different task runners, and chooses the
task runner based on the technology stack of the project as well as the
skills of their team.

In this case, the choice is Robo, as the stack is in PHP and Robo has
fairly good support.

See https://medium.com/sitewards/moving-from-make-to-robo-7c7d53ac701c

== Design Notes ==

=== Docker Compose ===

This commit also includes an example task invoking a docker-compose
development environment. This environment will be added in future, but
serves to demonstrate Robo's for this commit.