Add Elasticsearch as an application dependency
The search data pool will be built using ElasticSearch. In order to
communicate with Elasticsearch some sort of PHP client should be
required, or the protocol communicated directly over a socket.

Elasticsearch themselves provide a PHP client for the most recent
version of Elasticsearch, and it stands to reason that it's a good one
to start implementation with.