Add Elastiquent configuration, product model
In order to get the application up and running Elastiquaint requires a
model in order to populate data.

This commit adds this model, as well as additional configuration
Elastiquaint requires to connect to the ES database.

Future work will make this configuration expressed via the environment,
as well as add configuration to seed the database with appropriate
product values.