Allow persisting the product model
This commit takes the first step that allows persisting the product
model into an index that has already been created within ElasticSearch.

== Design Notes ==

=== Inheritance ===

A large inheritance tree is starting to emerge for the entities. The
code is still "reasonable", and the responsibilities suitably different
but the large inheritance tree may indicate that the classes are
becoming too complex, or that they need recompositing