Allow product seeder to run on new databases
Currently the product seeder is the primary way in which products are
injected into the database for testing.

However, there is a bug in which the seeder will only work if an
elasticsearch index already exists, and the seeder is required to delete
that before repopulating the index.

This commit addresses that by attempting to delete the previous index,
but ignoring the failure of that deletion in the case that the index has
already been deleted.