Add middleware that validates the request
== Design Notes ==

=== Not using the built in middleware ===

The built in middleware provided by the application appears to make the
assumption that the response will in some way have a viewable element,
or otherwise tolerate the "flash messages" style approach.

This is not the design of this JSON only API. Rather, this API is
designed to return an extremely strict set of well formed responses.

Accordingly, the validation provided by default was eschewed in favour
of a custom implementation of validation.