Consume the query parameter in the API
In previous work, the Elasticsearch API was queried within the API but
not factoring in the query parameter -- essentially returning the same,
fixed list of entities.

This commit factors the query parameter into the API, tailoring the
search results based on what the query parameter says.

The query parameter is simply fed directly in lucene syntax to
ElasticSearch.