Add Keycloak as an identity provider
A microservice needs an extremely limited set of responsibilities. In
this case, the service should not manage its own user authentication,
but rather delegate authentication to a third party service via
standard protocols.

This commit adds such a service, called "Keycloak". This service is from
Red Hat, and is designed to do much of the heavy lifting of user account
management.

== Design Notes ==

=== Docker Compose ===

While developing the SSO tooling on top of keycloak it became apparent
that there needs to be consistent domain names both internal and
external to the docker-compose network.

By naming the containers after FQDNs, they're discoverable locally at
that address as well as can be inserted into the host file.

This commit additionally adds some variables used by keycloak/oauth
while verifying the authentication.