Implement a layer of authentication
This commit builds in the previous work to include keycloak as the
canonical authentication server by implementing authentication based on
OpenID, and the keycloak implementation.

== Design Notes ==

=== Alternative Design: Manually check the JWT ===

While investigating this task the author learned a considerable amount
authentication, and managing authentication via Keycloak. The first pass
at this manually queried Keycloak for a given key, and manually handled
oAuth implementation. However, it turns out that OpenID builds on oAuth,
and suits the authors requirements nicely.

=== Restructuring the failure handler ===

While doing this work it became clear that we should not construct
responses in the middleware, even for error conditions.

Rather, the implementation of OpenID provided the good example of simply
dispatching exceptions in the middleware stack, and deferring handling
of those extensions to the router exception handler.

This commit takes advantage of this pattern to return an appropriate
JSON response in all classes of issue, including critical (but
recoverable) server failures, as well as the invariable authentication
failures.

=== Removal of auth middleware ===

This commit also removes the boilerplate authentication middleware, as
it is not going to be used.

=== App Service Provider ===

As part of this work limitations with lumens database implementation
became apparent. They were addressed in the app service provider as
implemented in other projects.

=== Helpers ===

All that can be said with the app service provider can also be said
about helpers.

=== Database, SQLite ===

The extension that was implemented (Fudarious\OIDConnect) did some
extended checking of the state parameter of the oAuth request /
response authentication workflow.

To do this, they need to retain the token in some sort of state. The
natural state to do this is either the KV store Redis, or the database.

This commit thus enables the database, but only as an SQLite
implementation. In the case that the database is lost, authentication
might fail but can be retried.

=== php-sqlite3 ===

This commit includes php-sqlite3 extension in the docker container as it
was not there previously.

=== Limitations ===

Currently the authentication doesn't factor in scope while determining
access to resources. At this time, only the fact a user has a session is
enough to garner them the required access.