Add Pagination to the API definition
This commit adds pagination to the API definition. This allows the
consumer to request only a limited subset of the results for processing,
rather than being forced to consume all results at a single point no
matter the length.

== Design Notes ==

=== Parameter names ===

The names for the pagination parameters have been drawn from:

  https://cloud.google.com/apis/design/design_patterns#list_pagination

Though those docs are design for Protobuf, Google has a tremendous
amount of wisdom about API design and it stands to reason they've had to
deal with pagination before.