Implement a collection entity for the product object.
In order to reason about large pools of the object there is another
object type, the "collection".

This commit adds the collection object.

The git notes are terrible here. Future work should not do this.

== Design Notes ==

=== General Design Notes ===

This was implemented at great speed, and without much thought. It thus
has a generally "hurried" quality, with the concepts not well thought
out, tests not written and the code generally not super clean.

It's just a conversation piece at this point.

=== Composition over inheritance ===

Normally it would be better to create a set of abstractions that could
be expressed together such they can be composed into a logical unit,
rather than a single logical unit that shares parents.

However, due to time constraints and the pressure associated with the
interview process this commit instead proposes an inheritance model.

Future work should consider these abstractions in more depth such that
they can be more accurately composed together

=== Single type of entity ===

This commit by inference limits the entity types to the "entity" type,
rather than allowing the collection to be used for some reasonable
object rather then the "entity" specified.

This limits the usability of this object, but the service more generally
has a very refined responsibility of "search", and thus should not
cause ongoing maintenance issues.

=== Collection object is responsible for search behaviour ===

The collection object includes the responsibility to transform the
arbitrary search query into some sort of queryable object.

Normally it would be better to abstract this transformation out into
a set of characteristics the collection can query for however that
would increase the amount of work required to fulfill requirements
beyond time constraints.

=== null properties as default for class properties ===

Within strictly typed languages (go), defaults are generally a "nullable"
or "falsy" value associated with the entity. However, in this case the
implementation is in PHP, and there is a type that can be interpreted
as "unset".

This is used here to indicate the properties are unmodified from their
original state.

=== Filter is limited to simple values ===

At this time, there are several shortfalls with the filter process.
These include:

- The ability to only filter by simple , key → value pairs
- The inability to filter by multiple properties
- The inability to filter by wildcards

However, as a first pass as this functionality simple key → value
pairs seem fine.

=== A general trend toward immutability ===

One of the design decisions made during this implementation was to not
implement a method to determine the state of the collection; whether it
had yet queried the data store.

This, in the authors mind, breaks a certain level of encapsulation. It
is difficult to provide the guarantees of a point in time nature of the
data, and exposing a method that allows whether the collection is loaded
(and cached) allows implementers of this class to do a proxy check, and
asks for the method that allows the class to be "reloaded".

By not implementing this method we force users to reconstruct the object,
thus allowing users to consistently guarantee "just loaded" semantics.

=== Using setters instead of constructor arguments ===

This object is somewhat "lazy initialised", which make it more difficult
to reason about the objects state.

It would be better to swap out these setters for constructor arguments,
further pushing towards immutability.

However, due to time constraints setters were simply easier.